Fix completion sorting in remote method calls
