+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/completions/providers/AbstractCompletionProvider.java
@@ -747,4 +747,25 @@ public abstract class AbstractCompletionProvider<T extends Node> implements Ball
         
         return Optional.of(new SnippetCompletionItem(context, snippetBlock));
     }
+
+    /**
+     * Sets the sort text of the provided completion item based on provided rank and context type.
+     *
+     * @param context        Completion context
+     * @param completionItem Completion item
+     * @param rank           A secondary rank to be considered other than assignability
+     */
+    protected void sortByAssignability(BallerinaCompletionContext context, LSCompletionItem completionItem, int rank) {
+        Optional<TypeSymbol> contextType = context.getContextType();
+        String sortText = "";
+        // First we sort the assignable items above others, and then sort by the rank
+        if (contextType.isPresent() && SortingUtil.isCompletionItemAssignable(completionItem, contextType.get())) {
+            sortText += SortingUtil.genSortText(1);
+        } else {
+            sortText += SortingUtil.genSortText(2);
+        }
+        sortText += SortingUtil.genSortText(rank);
+
+        completionItem.getCompletionItem().setSortText(sortText);
+    }
 }
